Crate deskulpt_widgets

This crate implements internal APIs for Deskulpt widgets management.

âš ï¸ This crate is meant to be used internally by the Deskulpt application and is not designed to support plugin authors or other users directly. Private items are documented for reference of Deskulpt developers.

Modules

catalog ğŸ”’

Deskulpt widget descriptors and catalog.

commands ğŸ”’

Tauri commands.

events ğŸ”’

Tauri events.

manager ğŸ”’

Deskulpt widgets manager and its APIs.

manifest ğŸ”’

Manifest files that describe Deskulpt widgets.

render ğŸ”’

Bundling and rendering of Deskulpt widgets.

setup ğŸ”’

Setup state of frontend windows.

Structs

WidgetsManager

Manager for Deskulpt widgets.

Traits

WidgetsExt ğŸ”’

Extension to Manager for accessing Deskulpt widgets APIs.

Functions

init

Initialize the internal Deskulpt widgets plugin.