deskulpt_widgets::manifest::load

Trait LoadManifest

pub trait LoadManifest: Sized + DeserializeOwned {
    const FILE_NAME: &'static str;

    // Provided method
    fn load(dir: &Path) -> Result<Option<Self>> { ... }
}

Helper trait for loading manifest files from a directory.

Required Associated Constants

const FILE_NAME: &'static str

The name of the manifest file to look for in the directory.

Provided Methods

fn load(dir: &Path) -> Result<Option<Self>>

Load the manifest file from the given directory.

This method does not treat the absence of the manifest file as an error. Instead, it returns Ok(None) in that case. If the file exists but cannot be read or parsed, an error is returned. Otherwise, the parsed manifest is returned wrapped in Ok(Some(...)).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.